rename PrayingExperienceConnected to PrayerExperienceConnected
- updates imports and references
